requirements.txt
sanic==22.6.2
accelerate
timm
pillow
torch
torchvision
